Consolidate Example: simplify_udwf_expression.rs into advanced_udwf.rs
